+++ Tests/GeometriaTests/P5Printer.swift
@@ -37,15 +37,22 @@ class P5Printer {
     // MARK: - Geometry
 
     private func _vertexRadius<Scalar: Numeric>() -> Scalar {
-        return 2
+        return 4
     }
     
     func add<Vector: Vector2Type>(_ point: Vector, style: Style? = nil) where Vector.Scalar: Numeric & CustomStringConvertible {
-        add(Circle2(center: point, radius: _vertexRadius()), style: style)
+        let circle = Circle2(center: point, radius: _vertexRadius())
+
+        addStyleSet(style ?? styling.geometry)
+        addDrawLine("circle(\(vec2String(circle.center)), \(circle.radius) / renderScale)")
     }
     
     func add<Vector: Vector3Type>(_ point: Vector, style: Style? = nil) where Vector.Scalar: Numeric & CustomStringConvertible {
-        add(Sphere3(center: point, radius: _vertexRadius()), style: style)
+        let sphere = Sphere3(center: point, radius: _vertexRadius())
+
+        is3D = true
+        
+        addDrawLine(sphere3String_customRadius(sphere, radius: "\(sphere.radius) / renderScale"))
     }
     
     func add<V: Vector2Type>(_ ellipse: Ellipsoid<V>, style: Style? = nil) where V.Scalar: CustomStringConvertible {
@@ -140,15 +147,15 @@ class P5Printer {
         addDrawLine("circle(\(vec2String(circle.center)), \(circle.radius))")
     }
     
-    func add<V: Vector2Additive & VectorDivisible>(_ aabb: AABB2<V>, style: Style? = nil) where V.Scalar: CustomStringConvertible {
-        addStyleSet(style ?? styling.geometry)
-        addDrawLine("rect(\(vec2String(aabb.minimum)), \(vec2String(aabb.maximum)))")
-    }
-    
     func add<V: Vector3Type>(_ sphere: Sphere3<V>, style: Style? = nil) where V.Scalar: Numeric & CustomStringConvertible {
         is3D = true
         
-        addDrawLine("drawSphere(\(vec3String(sphere.center)), \(sphere.radius))")
+        addDrawLine(sphere3String(sphere))
+    }
+    
+    func add<V: Vector2Additive & VectorDivisible>(_ aabb: AABB2<V>, style: Style? = nil) where V.Scalar: CustomStringConvertible {
+        addStyleSet(style ?? styling.geometry)
+        addDrawLine("rect(\(vec2String(aabb.minimum)), \(vec2String(aabb.maximum)))")
     }
     
     func add<V: Vector3Additive & VectorDivisible>(_ aabb: AABB3<V>, style: Style? = nil) where V.Scalar: Numeric & CustomStringConvertible {
@@ -158,16 +165,6 @@ class P5Printer {
         addDrawLine("push()")
         addDrawLine("translate(\(vec3String(aabb.minimum + aabb.size / 2)))")
         addDrawLine("box(\(vec3String(aabb.size)))")
-        addDrawLine("pop()")
-    }
-    
-    func add<R: RectangleType>(_ rectangle: R, style: Style? = nil) where R.Vector: Vector2Additive & VectorDivisible, R.Vector.Scalar: Numeric & CustomStringConvertible {
-        is3D = true
-        
-        addStyleSet(style ?? styling.geometry)
-        addDrawLine("push()")
-        addDrawLine("translate(\(vec2String(rectangle.location + rectangle.size / 2)))")
-        addDrawLine("box(\(vec2String(rectangle.size)))")
         addDrawLine("pop()")
     }
     
@@ -260,14 +257,19 @@ class P5Printer {
             printKeyPressed()
         }
         
-        if drawGrid {
+        if drawGrid && !is3D {
             printLine("")
             printDrawGrid2D()
         }
         
-        if drawOrigin && is3D {
+        if drawOrigin {
             printLine("")
-            printDrawOrigin3D()
+
+            if is3D {
+                printDrawOrigin3D()
+            } else {
+                printDrawOrigin2D()
+            }
         }
         
         if shouldPrintDrawNormal {
@@ -334,7 +336,7 @@ class P5Printer {
     }
     
     func addStrokeWeightSet(_ value: String) {
-        let line = "strokeWeight(\(value))"
+        let line = "strokeWeight(\(value) / sceneScale)"
 
         if _lastStrokeWeightCall == line { return }
         
@@ -399,7 +401,7 @@ class P5Printer {
             printLine("ellipseMode(RADIUS)")
             printLine("rectMode(CORNERS)")
 
-            if shouldStartDebugMode && is3D {
+            if is3D && shouldStartDebugMode {
                 printLine("debugMode(GRID)")
             }
             
@@ -425,11 +427,8 @@ class P5Printer {
             printLine("")
             printLine("strokeWeight(3 / sceneScale)")
 
-            if drawGrid {
-                printLine("drawGrid()")
-            }
             if drawOrigin && is3D {
-                printLine("drawOrigin()")
+                printLine("drawOrigin3D()")
             }
 
             if is3D {
@@ -438,6 +437,14 @@ class P5Printer {
 
             printLine("scale(renderScale)")
             
+            if drawGrid && !is3D {
+                printLine("")
+                printLine("drawGrid()")
+            }
+            if drawOrigin && !is3D {
+                printLine("drawOrigin2D()")
+            }
+
             printLine("")
             
             for draw in draws {
@@ -463,22 +470,35 @@ class P5Printer {
     
     func printDrawGrid2D() {
         indentedBlock("function drawGrid() {") {
-            printLine("stroke(0, 0, 0, 30)")
+            printLine("strokeWeight(1 / sceneScale)")
+            printLine("stroke(0, 0, 0, 20)")
             printLine("line(0, -20, 0, 20)")
             printLine("line(-20, 0, 20, 0)")
-            indentedBlock("for (var x = -10; x < 10; x++) {") {
-                printLine("stroke(0, 0, 0, 20)")
+            indentedBlock("for (var x = -20; x < 20; x++) {") {
                 printLine("line(x, -20, x, 20)")
             }
-            indentedBlock("for (var y = -10; y < 10; y++) {") {
-                printLine("stroke(0, 0, 0, 20)")
+            indentedBlock("for (var y = -20; y < 20; y++) {") {
                 printLine("line(-20, y, 20, y)")
             }
         }
     }
     
+    func printDrawOrigin2D() {
+        indentedBlock("function drawOrigin2D() {") {
+            let length: Double = 25.0
+            
+            printLine("strokeWeight(1 / sceneScale)")
+            printLine("// X axis")
+            printLine("stroke(255, 0, 0)")
+            printLine("line(0.0, 0.0, \(length) / renderScale, 0.0)")
+            printLine("// Y axis")
+            printLine("stroke(0, 255, 0)")
+            printLine("line(0.0, 0.0, 0.0, \(length) / renderScale)")
+        }
+    }
+    
     func printDrawOrigin3D() {
-        indentedBlock("function drawOrigin() {") {
+        indentedBlock("function drawOrigin3D() {") {
             let length: Double = 100.0
             
             let vx = Vector3D.unitX * length
@@ -566,6 +586,14 @@ class P5Printer {
     
     func vec2String<Vector: Vector2Type>(_ vec: Vector) -> String where Vector.Scalar: CustomStringConvertible {
         "\(vec.x), \(vec.y)"
+    }
+
+    func sphere3String<Vector: Vector3Type>(_ sphere: Sphere3<Vector>) -> String where Vector.Scalar: CustomStringConvertible {
+        "drawSphere(\(vec3String(sphere.center)), \(sphere.radius))"
+    }
+
+    func sphere3String_customRadius<Vector: Vector3Type>(_ sphere: Sphere3<Vector>, radius: String) -> String where Vector.Scalar: CustomStringConvertible {
+        "drawSphere(\(vec3String(sphere.center)), \(radius))"
     }
 
     func applyMatrix3DString<Scalar>(_ matrix: RotationMatrix3<Scalar>) -> [String] where Scalar: CustomStringConvertible {
@@ -720,7 +748,7 @@ extension P5Printer {
         var line: Style = Style(strokeColor: .black, strokeWeight: 2.0)
         var normalLine: Style = Style(strokeColor: .red.translucent, strokeWeight: 2.0)
         var tangentLine: Style = Style(strokeColor: .purple.translucent, strokeWeight: 2.0)
-        var geometry: Style = Style(strokeColor: .black)
+        var geometry: Style = Style(strokeColor: .black, strokeWeight: 2.0)
     }
 }