+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Adding `GeometriaAlgorithms` library that will feature useful computational geometry algorithm implementations.
     - Adding `SpatialTree<T>` type that encodes a simple [quad tree](https://en.wikipedia.org/wiki/Quadtree)/[octree](https://en.wikipedia.org/wiki/Octree) structures.
 
+- Added `DivisibleRectangleType.subdivided()`.
+
 - Fixing `NSphere.project(_:Vector)` being offset from correct point by `-self.center`.
 
 - `TakeVector3` has been removed in favor of `VectorTakeable` protocol.